feat : FeedFragment 리팩토링 feat : PreferencesManager에서 userId 가져오는 거 companion object에서 가져오면됨 feat : 좋아요 thumbup 하나로 구현할 수 있게함

+++ AOS/app/src/main/java/com/omnyom/yumyum/model/like/LikeRequest.kt
@@ -1,7 +1,7 @@
 package com.omnyom.yumyum.model.like
 
-class LikeRequest (private val feedId: Int, private val userId: Int) {
-    fun get() : HashMap<String, Int> {
+class LikeRequest (private val feedId: Long, private val userId: Long) {
+    fun get() : HashMap<String, Long> {
         return hashMapOf(
                 "feedId" to feedId,
                 "userId" to userId
